Enhance throws of a match with player names
Frontend implementations of this feature are currently not working or hacky. So instead the backend does the work and injects the player name into every throw when the throws attribute is accessed.

Once this stops being a viable solution due to performance problems we can start storing the name permanently, implement a client side solution or think about a different data model.